+++ src/main/java/com/clone/ohouse/error/order/OrderError.java
@@ -1,4 +1,4 @@
-package com.clone.ohouse.store.error.order;
+package com.clone.ohouse.error.order;
