+++ src/main/java/com/clone/ohouse/store/domain/OrderService.java
@@ -11,8 +11,8 @@ import com.clone.ohouse.store.domain.product.ProductRepository;
 import com.clone.ohouse.store.domain.product.Product;
 import com.clone.ohouse.store.domain.storeposts.StorePosts;
 import com.clone.ohouse.store.domain.storeposts.StorePostsRepository;
-import com.clone.ohouse.store.error.order.OrderError;
-import com.clone.ohouse.store.error.order.OrderFailException;
+import com.clone.ohouse.error.order.OrderError;
+import com.clone.ohouse.error.order.OrderFailException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
