+++ src/main/java/com/clone/ohouse/store/domain/PaymentService.java
@@ -6,10 +6,10 @@ import com.clone.ohouse.store.domain.payment.Payment;
 import com.clone.ohouse.store.domain.payment.PaymentRepository;
 import com.clone.ohouse.store.domain.payment.PaymentResultStatus;
 import com.clone.ohouse.store.domain.payment.dto.*;
-import com.clone.ohouse.store.error.order.OrderError;
-import com.clone.ohouse.store.error.order.OrderFailException;
-import com.clone.ohouse.store.error.order.PaymentError;
-import com.clone.ohouse.store.error.order.PaymentFailException;
+import com.clone.ohouse.error.order.OrderError;
+import com.clone.ohouse.error.order.OrderFailException;
+import com.clone.ohouse.error.order.PaymentError;
+import com.clone.ohouse.error.order.PaymentFailException;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
