+++ src/main/java/com/clone/ohouse/store/domain/product/Product.java
@@ -4,8 +4,8 @@ import com.clone.ohouse.store.domain.item.Item;
 import com.clone.ohouse.store.domain.storeposts.StorePosts;
 import com.clone.ohouse.store.domain.order.Order;
 import com.clone.ohouse.store.domain.order.OrderedProduct;
-import com.clone.ohouse.store.error.order.OrderError;
-import com.clone.ohouse.store.error.order.OrderFailException;
+import com.clone.ohouse.error.order.OrderError;
+import com.clone.ohouse.error.order.OrderFailException;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
